+++ htdocs/components/RequestDashboardComponent/subComponents/SalesRequestComponent.js
@@ -106,6 +106,20 @@ class SalesRequestComponent extends MainComponent {
                 }, problem.problemID)
             },
             {
+                path: "linkedSalesOrderID",
+                label: "",
+                hdToolTip: "Sales Order",
+                hdClassName: "text-center",
+                icon: "fal fa-2x fa-tag color-gray2 pointer",
+                className: "text-center",
+                classNameColumn: "",
+                sortable: false,
+                content: (problem) => el('a', {
+                    href: `SalesOrder.php?action=displaySalesOrder&ordheadID=${problem.linkedSalesOrderID}`,
+                    target: '_blank'
+                }, problem.linkedSalesOrderID)
+            },
+            {
                 path: "requestBody",
                 label: "",
                 key: "requestBody",
@@ -292,4 +306,3 @@ class SalesRequestComponent extends MainComponent {
 
 export default SalesRequestComponent;
 
-